Refactor: create two files scrape.py and product.py
